docker/app/requirements.txt
anyio==3.6.2
boto3==1.26.133
botocore==1.29.138
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.85.2
filelock==3.10.2
fonttools==4.39.4
ftfy==6.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.7
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.19.0
wcwidth==0.2.6
zipp==3.15.0
redis==4.5.5
ml2rt==0.2.0
redisai==1.3.0